requirements.txt
flask
annotated-types==0.6.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==23.1.0
Automat==22.10.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/b/abs_0agyz1wsr4/croot/beautifulsoup4-split_1681493048687/work
bert-extractive-summarizer==0.10.1
blinker==1.7.0
blis==0.7.11
Bottleneck @ file:///C:/ci_311/bottleneck_1676500016583/work
brotlipy==0.7.0
catalogue==2.0.10
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.16.0
cloudpickle @ file:///C:/b/abs_3796yxesic/croot/cloudpickle_1683040098851/work
colorama @ file:///C:/ci_311/colorama_1676422310965/work
comm @ file:///C:/ci_311/comm_1678376562840/work
confection==0.1.4
constantly==23.10.4
contourpy @ file:///C:/ci_311/contourpy_1676431756017/work
cryptography @ file:///C:/b/abs_13590mi9q9/croot/cryptography_1689373706078/work
css-html-js-minify==2.5.5
cssselect==1.2.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cymem==2.0.8
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dominate==2.9.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.13.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
fonttools==4.25.0
fsspec==2023.12.2
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.20.1
hyperlink==21.0.0
idna @ file:///C:/ci_311/idna_1676424932545/work
incremental==22.10.0
ipykernel @ file:///C:/b/abs_07rkft_vaz/croot/ipykernel_1691121700587/work
ipython @ file:///C:/b/abs_254uk73z5b/croot/ipython_1691532131313/work
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi @ file:///C:/ci_311/jedi_1679427407646/work
Jinja2==3.1.2
jmespath==1.0.1
joblib @ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
jupyter_client @ file:///C:/b/abs_059idvdagk/croot/jupyter_client_1680171872444/work
jupyter_core @ file:///C:/b/abs_9d0ttho3bs/croot/jupyter_core_1679906581955/work
kiwisolver @ file:///C:/ci_311/kiwisolver_1676431979301/work
langcodes==3.3.0
lorem==0.1.1
lxml==4.9.3
m3u8==4.0.0
MarkupSafe==2.1.3
matplotlib @ file:///C:/b/abs_49b2acwxd4/croot/matplotlib-suite_1679593486357/work
matplotlib-inline @ file:///C:/ci_311/matplotlib-inline_1676425798036/work
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
murmurhash==1.0.10
Naked==0.1.32
nest-asyncio @ file:///C:/ci_311/nest-asyncio_1676423519896/work
networkx==3.2.1
nltk==3.8.1
numexpr @ file:///C:/b/abs_afm0oewmmt/croot/numexpr_1683221839116/work
numpy @ file:///C:/b/abs_f6napi3n6e/croot/numpy_and_numpy_base_1691091651337/work
opencv-python==4.8.1.78
openpyxl==3.0.10
outcome==1.2.0
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1692298018988/work
parsel==1.8.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
ply==3.11
preshed==3.0.9
prompt-toolkit @ file:///C:/ci_311/prompt-toolkit_1676425940920/work
Protego==0.3.0
psutil @ file:///C:/ci_311_rebuilds/psutil_1679005906571/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pydantic==2.5.3
pydantic_core==2.14.6
PyDispatcher==2.0.7
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing @ file:///C:/ci_311/pyparsing_1678502182533/work
pyphen==0.14.0
PyQt5==5.15.7
PyQt5-sip @ file:///C:/ci_311/pyqt-split_1676428895938/work/pyqt_sip
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytube==15.0.0
pytz @ file:///C:/ci_311/pytz_1676427070848/work
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
queuelib==1.6.2
regex==2023.12.25
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
requests-file==1.5.1
safetensors==0.4.1
scikit-learn @ file:///C:/b/abs_38k7ridbgr/croot/scikit-learn_1684954723009/work
scipy==1.11.1
Scrapy==2.11.0
selenium==4.12.0
service-identity==23.1.0
shellescape==3.8.1
sip @ file:///C:/ci_311/sip_1676427825172/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve @ file:///C:/b/abs_a989exj3q6/croot/soupsieve_1680518492466/work
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spyder-kernels @ file:///C:/b/abs_e788a8_4y9/croot/spyder-kernels_1691599588437/work
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
textstat==0.7.3
thinc==8.2.2
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tldextract==5.1.1
tokenizers==0.15.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==2.1.2
tornado @ file:///C:/b/abs_61jhmrrua1/croot/tornado_1690848767317/work
tqdm==4.66.1
traitlets @ file:///C:/ci_311/traitlets_1676423290727/work
transformers==4.36.2
trio==0.22.2
trio-websocket==0.10.4
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typer==0.9.0
typing_extensions==4.9.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
visitor==0.1.3
w3lib==2.1.2
wasabi==1.1.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weasel==0.3.4
webdriver-manager==4.0.1
Werkzeug==3.0.1
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
wsproto==1.2.0
zope.interface==6.1
